Make Sentinels check that GestureEvent is within chart area
